+++ src/main/java/fr/insee/genesis/controller/rest/responses/ResponseController.java
@@ -1,4 +1,4 @@
-package fr.insee.genesis.controller.rest;
+package fr.insee.genesis.controller.rest.responses;
 
 import fr.insee.bpm.exceptions.MetadataParserException;
 import fr.insee.bpm.metadata.model.VariablesMap;
@@ -6,8 +6,6 @@ import fr.insee.bpm.metadata.reader.ddi.DDIReader;
 import fr.insee.bpm.metadata.reader.lunatic.LunaticReader;
 import fr.insee.genesis.Constants;
 import fr.insee.genesis.controller.adapter.LunaticXmlAdapter;
-import fr.insee.genesis.controller.dto.CampaignWithQuestionnaire;
-import fr.insee.genesis.controller.dto.QuestionnaireWithCampaign;
 import fr.insee.genesis.controller.dto.SurveyUnitDto;
 import fr.insee.genesis.controller.dto.SurveyUnitId;
 import fr.insee.genesis.controller.dto.SurveyUnitSimplified;
@@ -54,7 +52,6 @@ import java.nio.file.Paths;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Set;
 import java.util.stream.Stream;
 
 @RequestMapping(path = "/responses" )
@@ -181,7 +178,7 @@ public class ResponseController {
 
     //GET
     @Operation(summary = "Retrieve responses for an interrogation, using IdUE and IdQuestionnaire from Genesis Database")
-    @GetMapping(path = "/get-responses/by-ue-and-questionnaire")
+    @GetMapping(path = "/by-ue-and-questionnaire")
     public ResponseEntity<List<SurveyUnitModel>> findResponsesByUEAndQuestionnaire(@RequestParam("idUE") String idUE,
                                                                                    @RequestParam("idQuestionnaire") String idQuestionnaire) {
         List<SurveyUnitModel> responses = surveyUnitService.findByIdsUEAndQuestionnaire(idUE, idQuestionnaire);
@@ -189,7 +186,7 @@ public class ResponseController {
     }
 
     @Operation(summary = "Retrieve responses for an interrogation, using IdUE and IdQuestionnaire from Genesis Database with the latest value for each available state of every variable")
-    @GetMapping(path = "/get-responses/by-ue-and-questionnaire/latest-states")
+    @GetMapping(path = "/by-ue-and-questionnaire/latest-states")
     public ResponseEntity<SurveyUnitDto> findResponsesByUEAndQuestionnaireLatestStates(
             @RequestParam("idUE") String idUE,
             @RequestParam("idQuestionnaire") String idQuestionnaire) {
@@ -198,7 +195,7 @@ public class ResponseController {
     }
 
     @Operation(summary = "Retrieve all responses (for all interrogations) of one questionnaire")
-    @GetMapping(path = "/get-responses/by-questionnaire")
+    @GetMapping(path = "/by-questionnaire")
     public ResponseEntity<Path> findAllResponsesByQuestionnaire(@RequestParam("idQuestionnaire") String idQuestionnaire) {
         log.info("Try to find all responses of questionnaire : {}", idQuestionnaire);
 
@@ -220,7 +217,7 @@ public class ResponseController {
     }
 
     @Operation(summary = "Retrieve responses for an interrogation, using IdUE and IdQuestionnaire from Genesis Database. It returns only the latest value of each variable regardless of the state.")
-    @GetMapping(path = "/get-responses/by-ue-and-questionnaire/latest")
+    @GetMapping(path = "/by-ue-and-questionnaire/latest")
     public ResponseEntity<List<SurveyUnitModel>> getLatestByUE(@RequestParam("idUE") String idUE,
                                                                @RequestParam("idQuestionnaire") String idQuestionnaire) {
         List<SurveyUnitModel> responses = surveyUnitService.findLatestByIdAndByIdQuestionnaire(idUE, idQuestionnaire);
@@ -228,7 +225,7 @@ public class ResponseController {
     }
 
     @Operation(summary = "Retrieve responses for an interrogation, using IdUE and IdQuestionnaire from Genesis Database. For a given mode, it returns only the latest value of each variable regardless of the state. The result is one object by mode in the output")
-    @GetMapping(path = "/get-simplified-response/by-ue-questionnaire-and-mode/latest")
+    @GetMapping(path = "/simplified/by-ue-questionnaire-and-mode/latest")
     public ResponseEntity<SurveyUnitSimplified> getLatestByUEOneObject(@RequestParam("idUE") String idUE,
                                                                              @RequestParam("idQuestionnaire") String idQuestionnaire,
                                                                              @RequestParam("mode") Mode mode) {
@@ -252,7 +249,7 @@ public class ResponseController {
     @Operation(summary = "Retrieve all responses for a questionnaire and a list of UE",
             description = "Return the latest state for each variable for the given ids and a given questionnaire.<br>" +
                     "For a given id, the endpoint returns a document by collection mode (if there is more than one).")
-    @PostMapping(path = "/get-simplified-responses/by-list-ue-and-questionnaire/latest")
+    @PostMapping(path = "/simplified/by-list-ue-and-questionnaire/latest")
     public ResponseEntity<List<SurveyUnitSimplified>> getLatestForUEList(@RequestParam("idQuestionnaire") String idQuestionnaire,
                                                                                @RequestBody List<SurveyUnitId> idUEs) {
         List<SurveyUnitSimplified> results = new ArrayList<>();
@@ -279,64 +276,6 @@ public class ResponseController {
             });
         });
         return ResponseEntity.ok(results);
-    }
-
-    @Operation(summary = "Retrieve all IdUEs for a given questionnaire")
-    @GetMapping(path = "/get-idUEs/by-questionnaire")
-    public ResponseEntity<List<SurveyUnitId>> getAllIdUEsByQuestionnaire(@RequestParam("idQuestionnaire") String idQuestionnaire) {
-        List<SurveyUnitId> responses = surveyUnitService.findDistinctIdUEsByIdQuestionnaire(idQuestionnaire);
-        return ResponseEntity.ok(responses);
-    }
-
-    @Operation(summary = "List sources/modes used for a given questionnaire")
-    @GetMapping(path = "/get-modes/by-questionnaire")
-    public ResponseEntity<List<Mode>> getModesByQuestionnaire(@RequestParam("idQuestionnaire") String idQuestionnaire) {
-        List<Mode> modes = surveyUnitService.findModesByIdQuestionnaire(idQuestionnaire);
-        return ResponseEntity.ok(modes);
-    }
-
-    @Operation(summary = "List sources/modes used for a given campaign")
-    @GetMapping(path = "/get-modes/by-campaign")
-    public ResponseEntity<List<Mode>> getModesByCampaign(@RequestParam("idCampaign") String idCampaign) {
-        List<Mode> modes = surveyUnitService.findModesByIdCampaign(idCampaign);
-        return ResponseEntity.ok(modes);
-    }
-
-    @Operation(summary = "List questionnaires in database")
-    @GetMapping(path = "/get-questionnaires")
-    public ResponseEntity<Set<String>> getQuestionnaires() {
-        Set<String> questionnaires = surveyUnitService.findDistinctIdQuestionnaires();
-        return ResponseEntity.ok(questionnaires);
-    }
-
-
-    @Operation(summary = "List questionnaires in database with their campaigns")
-    @GetMapping(path = "/get-questionnaires/with-campaigns")
-    public ResponseEntity<List<QuestionnaireWithCampaign>> getQuestionnairesWithCampaigns() {
-        List<QuestionnaireWithCampaign> questionnaireWithCampaignList =
-                surveyUnitService.findQuestionnairesWithCampaigns();
-        return ResponseEntity.ok(questionnaireWithCampaignList);
-    }
-
-    @Operation(summary = "List questionnaires used for a given campaign")
-    @GetMapping(path = "/get-questionnaires/by-campaign")
-    public ResponseEntity<Set<String>> getQuestionnairesByCampaign(@RequestParam("idCampaign") String idCampaign) {
-        Set<String> questionnaires = surveyUnitService.findIdQuestionnairesByIdCampaign(idCampaign);
-        return ResponseEntity.ok(questionnaires);
-    }
-
-    @Operation(summary = "List campaigns in database")
-    @GetMapping(path = "/get-campaigns")
-    public ResponseEntity<Set<String>> getCampaigns() {
-        Set<String> campaigns = surveyUnitService.findDistinctIdCampaigns();
-        return ResponseEntity.ok(campaigns);
-    }
-
-    @Operation(summary = "List campaigns in database with their questionnaires")
-    @GetMapping(path = "/get-campaigns/with-questionnaires")
-    public ResponseEntity<List<CampaignWithQuestionnaire>> getCampaignsWithQuestionnaires() {
-        List<CampaignWithQuestionnaire> questionnairesByCampaigns = surveyUnitService.findCampaignsWithQuestionnaires();
-        return ResponseEntity.ok(questionnairesByCampaigns);
     }
 
     //Utilities
