--- src/test/java/fr/insee/genesis/controller/rest/ResponseControllerTest.java
@@ -1,629 +1,0 @@
-package fr.insee.genesis.controller.rest;
-
-import cucumber.TestConstants;
-import fr.insee.genesis.Constants;
-import fr.insee.genesis.controller.dto.CampaignWithQuestionnaire;
-import fr.insee.genesis.controller.dto.QuestionnaireWithCampaign;
-import fr.insee.genesis.controller.dto.SurveyUnitDto;
-import fr.insee.genesis.controller.dto.SurveyUnitId;
-import fr.insee.genesis.controller.dto.SurveyUnitSimplified;
-import fr.insee.genesis.controller.utils.ControllerUtils;
-import fr.insee.genesis.domain.model.surveyunit.CollectedVariable;
-import fr.insee.genesis.domain.model.surveyunit.DataState;
-import fr.insee.genesis.domain.model.surveyunit.Mode;
-import fr.insee.genesis.domain.model.surveyunit.SurveyUnitModel;
-import fr.insee.genesis.domain.model.surveyunit.Variable;
-import fr.insee.genesis.domain.ports.api.SurveyUnitApiPort;
-import fr.insee.genesis.domain.service.surveyunit.SurveyUnitQualityService;
-import fr.insee.genesis.domain.service.surveyunit.SurveyUnitService;
-import fr.insee.genesis.infrastructure.utils.FileUtils;
-import fr.insee.genesis.stubs.ConfigStub;
-import fr.insee.genesis.stubs.SurveyUnitPersistencePortStub;
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.springframework.http.ResponseEntity;
-import org.springframework.util.FileSystemUtils;
-
-import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.time.LocalDateTime;
-import java.time.Month;
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Set;
-
-class ResponseControllerTest {
-    //Given
-    static ResponseController responseControllerStatic;
-    static SurveyUnitPersistencePortStub surveyUnitPersistencePortStub;
-
-    static List<SurveyUnitId> surveyUnitIdList;
-    //Constants
-    static final String defaultIdUE = "TESTIDUE";
-    static final String defaultIdQuest = "TESTIDQUESTIONNAIRE";
-
-    @BeforeAll
-    static void init() {
-        surveyUnitPersistencePortStub = new SurveyUnitPersistencePortStub();
-        SurveyUnitApiPort surveyUnitApiPort = new SurveyUnitService(surveyUnitPersistencePortStub);
-
-        FileUtils fileUtils = new FileUtils(new ConfigStub());
-        responseControllerStatic = new ResponseController(
-                surveyUnitApiPort
-                , new SurveyUnitQualityService()
-                , fileUtils
-                , new ControllerUtils(fileUtils)
-        );
-
-        surveyUnitIdList = new ArrayList<>();
-        surveyUnitIdList.add(new SurveyUnitId(defaultIdUE));
-    }
-
-    @BeforeEach
-    void reset() throws IOException {
-        //MongoDB stub management
-        surveyUnitPersistencePortStub.getMongoStub().clear();
-
-        addAdditionalDtoToMongoStub("TESTIDCAMPAIGN", defaultIdQuest ,LocalDateTime.of(2023, 1, 1, 0, 0, 0),LocalDateTime.of(2024, 1, 1, 0, 0, 0));
-
-
-        //Test file management
-        //Clean DONE folder
-        Path testResourcesPath = Path.of(TestConstants.TEST_RESOURCES_DIRECTORY);
-        if (testResourcesPath.resolve("DONE").toFile().exists())
-            Files.walk(testResourcesPath.resolve("DONE"))
-                    .sorted(Comparator.reverseOrder())
-                    .map(Path::toFile)
-                    .forEach(File::delete);
-
-        //Recreate data files
-        //SAMPLETEST-PARADATA-v1
-        //Root
-        if (!testResourcesPath
-                .resolve("IN")
-                .resolve("WEB")
-                .resolve("SAMPLETEST-PARADATA-v1")
-                .resolve("data.complete.validated.STPDv1.20231122164209.xml")
-                .toFile().exists()
-        ){
-            Files.copy(
-                    testResourcesPath
-                            .resolve("IN")
-                            .resolve("WEB")
-                            .resolve("SAMPLETEST-PARADATA-v1")
-                            .resolve("reponse-platine")
-                            .resolve("data.complete.partial.STPDv1.20231122164209.xml")
-                    , testResourcesPath
-                            .resolve("IN")
-                            .resolve("WEB")
-                            .resolve("SAMPLETEST-PARADATA-v1")
-                            .resolve("data.complete.partial.STPDv1.20231122164209.xml")
-            );
-            Files.copy(
-                    testResourcesPath
-                            .resolve("IN")
-                            .resolve("WEB")
-                            .resolve("SAMPLETEST-PARADATA-v1")
-                            .resolve("reponse-platine")
-                            .resolve("data.complete.validated.STPDv1.20231122164209.xml")
-                    , testResourcesPath
-                            .resolve("IN")
-                            .resolve("WEB")
-                            .resolve("SAMPLETEST-PARADATA-v1")
-                            .resolve("data.complete.validated.STPDv1.20231122164209.xml")
-            );
-        }
-        //Differential data
-        if (!testResourcesPath
-                .resolve("IN")
-                .resolve("WEB")
-                .resolve("SAMPLETEST-PARADATA-v1")
-                .resolve("differential")
-                .resolve("data")
-                .resolve("data.complete.validated.STPDv1.20231122164209.xml")
-                .toFile().exists()
-        ) {
-            Files.copy(
-                    testResourcesPath
-                            .resolve("IN")
-                            .resolve("WEB")
-                            .resolve("SAMPLETEST-PARADATA-v1")
-                            .resolve("reponse-platine")
-                            .resolve("data.complete.partial.STPDv1.20231122164209.xml")
-                    , testResourcesPath
-                            .resolve("IN")
-                            .resolve("WEB")
-                            .resolve("SAMPLETEST-PARADATA-v1")
-                            .resolve("differential")
-                            .resolve("data")
-                            .resolve("data.complete.partial.STPDv1.20231122164209.xml")
-            );
-            Files.copy(
-                    testResourcesPath
-                            .resolve("IN")
-                            .resolve("WEB")
-                            .resolve("SAMPLETEST-PARADATA-v1")
-                            .resolve("reponse-platine")
-                            .resolve("data.complete.validated.STPDv1.20231122164209.xml")
-                    , testResourcesPath
-                            .resolve("IN")
-                            .resolve("WEB")
-                            .resolve("SAMPLETEST-PARADATA-v1")
-                            .resolve("differential")
-                            .resolve("data")
-                            .resolve("data.complete.validated.STPDv1.20231122164209.xml")
-            );
-        }
-        //SAMPLETEST-PARADATA-v2
-        if (!testResourcesPath
-                .resolve("IN")
-                .resolve("WEB")
-                .resolve("SAMPLETEST-PARADATA-v2")
-                .resolve("data.complete.validated.STPDv2.20231122164209.xml")
-                .toFile().exists()
-        ){
-            Files.copy(
-                    testResourcesPath
-                            .resolve("IN")
-                            .resolve("WEB")
-                            .resolve("SAMPLETEST-PARADATA-v2")
-                            .resolve("reponse-platine")
-                            .resolve("data.complete.partial.STPDv2.20231122164209.xml")
-                    , testResourcesPath
-                            .resolve("IN")
-                            .resolve("WEB")
-                            .resolve("SAMPLETEST-PARADATA-v2")
-                            .resolve("data.complete.partial.STPDv2.20231122164209.xml")
-            );
-            Files.copy(
-                    testResourcesPath
-                            .resolve("IN")
-                            .resolve("WEB")
-                            .resolve("SAMPLETEST-PARADATA-v2")
-                            .resolve("reponse-platine")
-                            .resolve("data.complete.validated.STPDv2.20231122164209.xml")
-                    , testResourcesPath
-                            .resolve("IN")
-                            .resolve("WEB")
-                            .resolve("SAMPLETEST-PARADATA-v2")
-                            .resolve("data.complete.validated.STPDv2.20231122164209.xml")
-            );
-        }
-        //SAMPLETEST-NO-COLLECTED
-        if (!testResourcesPath
-                .resolve("IN")
-                .resolve("WEB")
-                .resolve("SAMPLETEST-NO-COLLECTED")
-                .resolve("differential")
-                .resolve("data")
-                .resolve("data_diff_no_collected.xml")
-                .toFile().exists()
-        ){
-            Files.copy(
-                    testResourcesPath
-                            .resolve("IN")
-                            .resolve("WEB")
-                            .resolve("SAMPLETEST-NO-COLLECTED")
-                            .resolve("data_diff_no_collected.xml")
-                    , testResourcesPath
-                            .resolve("IN")
-                            .resolve("WEB")
-                            .resolve("SAMPLETEST-NO-COLLECTED")
-                            .resolve("differential")
-                            .resolve("data")
-                            .resolve("data_diff_no_collected.xml")
-            );
-        }
-    }
-
-
-    //When + Then
-    @Test
-    void saveResponseFromXMLFileTest() throws Exception {
-        responseControllerStatic.saveResponsesFromXmlFile(
-                Path.of(TestConstants.TEST_RESOURCES_DIRECTORY, "IN/WEB/SAMPLETEST-PARADATA-v1/reponse-platine/data.complete.validated.STPDv1.20231122164209.xml").toString()
-                , Path.of(TestConstants.TEST_RESOURCES_DIRECTORY, "specs/SAMPLETEST-PARADATA-v1/ddi-SAMPLETEST-PARADATA-v1.xml").toString()
-                , Mode.WEB
-                , true
-        );
-
-        Assertions.assertThat(surveyUnitPersistencePortStub.getMongoStub()).isNotEmpty();
-    }
-
-    @Test
-    void saveOneFileNoCollected_NoNullPointerException(){
-        Assertions.assertThatCode(() -> responseControllerStatic.saveResponsesFromXmlFile(
-                Path.of(TestConstants.TEST_RESOURCES_DIRECTORY, "IN/WEB/SAMPLETEST-NO-COLLECTED/differential/data/data_diff_no_collected.xml").toString()
-                , Path.of(TestConstants.TEST_RESOURCES_DIRECTORY, "specs/SAMPLETEST-NO-COLLECTED/WEB/ddi_response_simple.xml").toString()
-                , Mode.WEB
-                , true
-        )).doesNotThrowAnyException();
-
-    }
-
-    @Test
-    void saveResponsesFromXmlCampaignFolderTest() throws Exception {
-        responseControllerStatic.saveResponsesFromXmlCampaignFolder(
-                "SAMPLETEST-PARADATA-v1"
-                , Mode.WEB
-                , true
-        );
-
-        Assertions.assertThat(surveyUnitPersistencePortStub.getMongoStub()).isNotEmpty();
-    }
-
-    @Test
-    void saveResponsesFromXmlCampaignFolderTest_noData() throws Exception {
-        surveyUnitPersistencePortStub.getMongoStub().clear();
-
-        responseControllerStatic.saveResponsesFromXmlCampaignFolder(
-                "TESTNODATA"
-                , Mode.WEB
-                , true
-        );
-
-        Assertions.assertThat(surveyUnitPersistencePortStub.getMongoStub()).isEmpty();
-    }
-
-    @Test
-    void saveResponsesFromAllCampaignFoldersTests(){
-        surveyUnitPersistencePortStub.getMongoStub().clear();
-        responseControllerStatic.saveResponsesFromAllCampaignFolders();
-
-        Assertions.assertThat(surveyUnitPersistencePortStub.getMongoStub()).isNotEmpty();
-    }
-
-    @Test
-    void findResponsesByUEAndQuestionnaireTest() {
-        ResponseEntity<List<SurveyUnitModel>> response = responseControllerStatic.findResponsesByUEAndQuestionnaire(defaultIdUE, defaultIdQuest);
-
-        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
-        Assertions.assertThat(response.getBody()).isNotNull().isNotEmpty();
-        Assertions.assertThat(response.getBody().getFirst().getIdUE()).isEqualTo(defaultIdUE);
-        Assertions.assertThat(response.getBody().getFirst().getIdQuest()).isEqualTo(defaultIdQuest);
-    }
-
-    @Test
-    void findAllResponsesByQuestionnaireTest() {
-        Path path = Path.of(TestConstants.TEST_RESOURCES_DIRECTORY, "OUT", defaultIdQuest);
-        File dir = new File(String.valueOf(path));
-        FileSystemUtils.deleteRecursively(dir);
-
-        ResponseEntity<Path> response = responseControllerStatic.findAllResponsesByQuestionnaire(defaultIdQuest);
-
-        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
-        Assertions.assertThat(response.getBody()).isNotNull();
-        Assertions.assertThat(Files.exists(path)).isTrue();
-        File[] dir_contents = dir.listFiles();
-        Assertions.assertThat(dir_contents).hasSize(1);
-        Assertions.assertThat(dir_contents[0].length()).isPositive().isNotNull();
-        FileSystemUtils.deleteRecursively(dir);
-        dir.deleteOnExit();
-    }
-
-    @Test
-    void getAllResponsesByQuestionnaireTestSequential() throws IOException {
-        //Given
-        surveyUnitPersistencePortStub.getMongoStub().clear();
-
-        for (int i = 0; i < Constants.BATCH_SIZE + 2; i++) {
-            addAdditionalDtoToMongoStub("TESTIDCAMPAIGN", defaultIdUE + i,LocalDateTime.of(2023, 1, 1, 0, 0, 0),LocalDateTime.of(2024, 1, 1, 0, 0, 0));
-        }
-
-        //When
-        ResponseEntity<Path> response = responseControllerStatic.findAllResponsesByQuestionnaire(defaultIdQuest);
-
-        //Then
-        Assertions.assertThat(response).isNotNull();
-        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
-        Assertions.assertThat(response.getBody()).isNotNull();
-
-        Assertions.assertThat(response.getBody()).isNotNull();
-        Assertions.assertThat(response.getBody().toFile()).isNotNull().exists();
-
-        Files.deleteIfExists(response.getBody());
-    }
-
-    @Test
-    void getLatestByUETest() {
-        addAdditionalDtoToMongoStub();
-
-        ResponseEntity<List<SurveyUnitModel>> response = responseControllerStatic.getLatestByUE(defaultIdUE, defaultIdQuest);
-
-        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
-        Assertions.assertThat(response.getBody()).isNotNull().isNotEmpty();
-        Assertions.assertThat(response.getBody().getFirst().getIdUE()).isEqualTo(defaultIdUE);
-        Assertions.assertThat(response.getBody().getFirst().getIdQuest()).isEqualTo(defaultIdQuest);
-        Assertions.assertThat(response.getBody().getFirst().getFileDate()).hasMonth(Month.FEBRUARY);
-    }
-
-    @Test
-    void getLatestByUEOneObjectTest() {
-        ResponseEntity<SurveyUnitSimplified> response = responseControllerStatic.getLatestByUEOneObject(defaultIdUE, defaultIdQuest, Mode.WEB);
-
-        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
-        Assertions.assertThat(response.getBody()).isNotNull();
-        Assertions.assertThat(response.getBody().getIdUE()).isEqualTo(defaultIdUE);
-        Assertions.assertThat(response.getBody().getIdQuest()).isEqualTo(defaultIdQuest);
-    }
-
-    @Test
-    void getLatestForUEListTest() {
-        ResponseEntity<List<SurveyUnitSimplified>> response = responseControllerStatic.getLatestForUEList(defaultIdQuest, surveyUnitIdList);
-
-        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
-        Assertions.assertThat(response.getBody()).isNotNull().isNotEmpty();
-        Assertions.assertThat(response.getBody().getFirst().getIdUE()).isEqualTo(defaultIdUE);
-    }
-
-    @Test
-    void getAllIdUEsByQuestionnaireTest() {
-        ResponseEntity<List<SurveyUnitId>> response = responseControllerStatic.getAllIdUEsByQuestionnaire(defaultIdQuest);
-
-        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
-        Assertions.assertThat(response.getBody()).isNotNull().isNotEmpty();
-        Assertions.assertThat(response.getBody().getFirst().getIdUE()).isEqualTo(defaultIdUE);
-    }
-
-    @Test
-    void getModesByQuestionnaireTest() {
-        ResponseEntity<List<Mode>> response = responseControllerStatic.getModesByQuestionnaire(defaultIdQuest);
-
-        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
-        Assertions.assertThat(response.getBody()).isNotNull().isNotEmpty().hasSize(1);
-        Assertions.assertThat(response.getBody().getFirst()).isEqualTo(Mode.WEB);
-    }
-
-    @Test
-    void getModesByCampaignTest() {
-        ResponseEntity<List<Mode>> response = responseControllerStatic.getModesByCampaign("TESTIDCAMPAIGN");
-
-        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
-        Assertions.assertThat(response.getBody()).isNotNull().isNotEmpty().hasSize(1);
-        Assertions.assertThat(response.getBody().getFirst()).isEqualTo(Mode.WEB);
-    }
-
-    @Test
-    void getCampaignsTest() {
-        addAdditionalDtoToMongoStub("TESTCAMPAIGN2","TESTQUESTIONNAIRE2");
-
-        ResponseEntity<Set<String>> response = responseControllerStatic.getCampaigns();
-
-        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
-        Assertions.assertThat(response.getBody()).isNotNull().isNotEmpty().containsExactly(
-                "TESTIDCAMPAIGN","TESTCAMPAIGN2");
-    }
-
-    @Test
-    void getQuestionnairesTest() {
-        addAdditionalDtoToMongoStub("TESTQUESTIONNAIRE2");
-
-        ResponseEntity<Set<String>> response = responseControllerStatic.getQuestionnaires();
-
-        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
-        Assertions.assertThat(response.getBody()).isNotNull().isNotEmpty().containsExactly(
-                defaultIdQuest,"TESTQUESTIONNAIRE2");
-    }
-
-    @Test
-    void getQuestionnairesByCampaignTest() {
-        addAdditionalDtoToMongoStub("TESTQUESTIONNAIRE2");
-
-        ResponseEntity<Set<String>> response = responseControllerStatic.getQuestionnairesByCampaign("TESTIDCAMPAIGN");
-
-        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
-        Assertions.assertThat(response.getBody()).isNotNull().isNotEmpty().containsExactly(
-                defaultIdQuest,"TESTQUESTIONNAIRE2");
-    }
-
-    @Test
-    void getCampaignsWithQuestionnairesTest() {
-        addAdditionalDtoToMongoStub("TESTQUESTIONNAIRE2");
-        addAdditionalDtoToMongoStub("TESTCAMPAIGN2","TESTQUESTIONNAIRE2");
-
-        ResponseEntity<List<CampaignWithQuestionnaire>> response = responseControllerStatic.getCampaignsWithQuestionnaires();
-
-        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
-        Assertions.assertThat(response.getBody()).isNotNull().isNotEmpty();
-        Assertions.assertThat(response.getBody().stream().filter(campaignWithQuestionnaire ->
-                        campaignWithQuestionnaire.getIdCampaign().equals("TESTIDCAMPAIGN")
-                                || campaignWithQuestionnaire.getIdCampaign().equals("TESTCAMPAIGN2")
-        )).isNotNull().isNotEmpty().hasSize(2);
-
-        Assertions.assertThat(response.getBody().stream().filter(
-                campaignWithQuestionnaire -> campaignWithQuestionnaire.getIdCampaign().equals("TESTIDCAMPAIGN")
-        ).findFirst().get().getQuestionnaires()).containsExactly(defaultIdQuest, "TESTQUESTIONNAIRE2");
-
-        Assertions.assertThat(response.getBody().stream().filter(
-                campaignWithQuestionnaire -> campaignWithQuestionnaire.getIdCampaign().equals("TESTCAMPAIGN2")
-        ).findFirst().get().getQuestionnaires()).containsExactly("TESTQUESTIONNAIRE2");
-    }
-
-    @Test
-    void getQuestionnairesWithCampaignsTest() {
-        addAdditionalDtoToMongoStub("TESTQUESTIONNAIRE2");
-        addAdditionalDtoToMongoStub("TESTCAMPAIGN2","TESTQUESTIONNAIRE2");
-
-        ResponseEntity<List<QuestionnaireWithCampaign>> response = responseControllerStatic.getQuestionnairesWithCampaigns();
-
-        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
-
-        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
-        Assertions.assertThat(response.getBody()).isNotNull().isNotEmpty();
-        Assertions.assertThat(response.getBody().stream().filter(questionnaireWithCampaign ->
-                questionnaireWithCampaign.getIdQuestionnaire().equals(defaultIdQuest)
-                        || questionnaireWithCampaign.getIdQuestionnaire().equals("TESTQUESTIONNAIRE2")
-        )).isNotNull().isNotEmpty().hasSize(2);
-
-        Assertions.assertThat(response.getBody().stream().filter(
-                questionnaireWithCampaign -> questionnaireWithCampaign.getIdQuestionnaire().equals(defaultIdQuest)
-        ).findFirst().get().getCampaigns()).containsExactly("TESTIDCAMPAIGN");
-
-        Assertions.assertThat(response.getBody().stream().filter(
-                questionnaireWithCampaign -> questionnaireWithCampaign.getIdQuestionnaire().equals("TESTQUESTIONNAIRE2")
-        ).findFirst().get().getCampaigns()).containsExactly("TESTIDCAMPAIGN", "TESTCAMPAIGN2");
-    }
-
-    // Perret tests
-    @Test
-    void getPerretSurveyDataTest(){
-        //GIVEN
-        //Recent Collected already in stub
-        //Old Collected
-        addAdditionalDtoToMongoStub(DataState.COLLECTED,
-                "C OLD C", //<Collected/External> <NEW or OLD> <Collected/Edited>
-                "E OLD C",
-                LocalDateTime.of(1999,2,2,0,0,0),
-                LocalDateTime.of(1999,2,2,0,0,0)
-        );
-
-        //Recent Edited
-        addAdditionalDtoToMongoStub(DataState.EDITED,
-                "C NEW E",
-                "E NEW E",
-                LocalDateTime.of(2025,2,2,0,0,0),
-                LocalDateTime.of(2025,2,2,0,0,0)
-        );
-
-        //Old Edited
-        addAdditionalDtoToMongoStub(DataState.EDITED,
-                "C OLD E",
-                "E OLD E",
-                LocalDateTime.of(1999,2,2,0,0,0),
-                LocalDateTime.of(1999,2,2,0,0,0)
-        );
-
-
-        //WHEN
-        ResponseEntity<SurveyUnitDto> response = responseControllerStatic.findResponsesByUEAndQuestionnaireLatestStates(
-                defaultIdUE,
-                defaultIdQuest
-        );
-
-
-        //THEN
-        SurveyUnitDto surveyUnitDto = response.getBody();
-        Assertions.assertThat(surveyUnitDto).isNotNull();
-
-        Assertions.assertThat(surveyUnitDto.getSurveyUnitId()).isEqualTo(defaultIdUE);
-
-        Assertions.assertThat(surveyUnitDto.getCollectedVariables().getFirst().getVariableName())
-                .isEqualTo("TESTIDVAR");
-
-        Assertions.assertThat(surveyUnitDto.getCollectedVariables().getFirst().getVariableStateDtoList()
-                        .stream().filter(
-                                variableStatePerret -> variableStatePerret.getState().equals(DataState.COLLECTED)
-                        ).toList().getFirst().getValue())
-                .isEqualTo("V1");
-        Assertions.assertThat(surveyUnitDto.getCollectedVariables().getFirst().getVariableStateDtoList()
-                        .stream().filter(
-                                variableStatePerret -> variableStatePerret.getState().equals(DataState.EDITED)
-                        ).toList().getFirst().getValue())
-                .isEqualTo("C NEW E");
-        Assertions.assertThat(surveyUnitDto.getCollectedVariables().getFirst().getVariableStateDtoList()
-                        .stream().filter(
-                                variableStatePerret -> variableStatePerret.getState().equals(DataState.COLLECTED)
-                        ).toList().getFirst().isActive())
-                .isFalse();
-        Assertions.assertThat(surveyUnitDto.getCollectedVariables().getFirst().getVariableStateDtoList()
-                        .stream().filter(
-                                variableStatePerret -> variableStatePerret.getState().equals(DataState.EDITED)
-                        ).toList().getFirst().isActive())
-                .isTrue();
-
-        Assertions.assertThat(surveyUnitDto.getExternalVariables().getFirst().getVariableName())
-                .isEqualTo("TESTIDVAR");
-        Assertions.assertThat(surveyUnitDto.getExternalVariables().getFirst().getVariableStateDtoList()
-                        .stream().filter(
-                                variableStatePerret -> variableStatePerret.getState().equals(DataState.COLLECTED)
-                        ).toList().getFirst().getValue())
-                .isEqualTo("V1");
-        Assertions.assertThat(surveyUnitDto.getExternalVariables().getFirst().getVariableStateDtoList()
-                        .stream().filter(
-                                variableStatePerret -> variableStatePerret.getState().equals(DataState.EDITED)
-                        ).toList().getFirst().getValue())
-                .isEqualTo("E NEW E");
-        Assertions.assertThat(surveyUnitDto.getExternalVariables().getFirst().getVariableStateDtoList()
-                        .stream().filter(
-                                variableStatePerret -> variableStatePerret.getState().equals(DataState.COLLECTED)
-                        ).toList().getFirst().isActive())
-                .isFalse();
-        Assertions.assertThat(surveyUnitDto.getExternalVariables().getFirst().getVariableStateDtoList()
-                        .stream().filter(
-                                variableStatePerret -> variableStatePerret.getState().equals(DataState.EDITED)
-                        ).toList().getFirst().isActive())
-                .isTrue();
-    }
-
-
-
-    // Utilities
-    private void addAdditionalDtoToMongoStub() {
-        addAdditionalDtoToMongoStub(defaultIdQuest);
-    }
-
-    private void addAdditionalDtoToMongoStub(String idQuestionnaire) {
-        addAdditionalDtoToMongoStub("TESTIDCAMPAIGN",idQuestionnaire);
-    }
-
-    private void addAdditionalDtoToMongoStub(String idCampaign, String idQuestionnaire) {
-        addAdditionalDtoToMongoStub(idCampaign,idQuestionnaire,LocalDateTime.of(2023, 2, 2, 0, 0, 0),LocalDateTime.of(2024, 2, 2, 0, 0, 0));
-    }
-
-    private void addAdditionalDtoToMongoStub(String idCampaign, String idQuestionnaire, LocalDateTime fileDate, LocalDateTime recordDate) {
-        List<Variable> externalVariableList = new ArrayList<>();
-        Variable variable = Variable.builder().idVar("TESTIDVAR").values(List.of(new String[]{"V1", "V2"})).build();
-        externalVariableList.add(variable);
-
-        List<CollectedVariable> collectedVariableList = new ArrayList<>();
-        CollectedVariable collectedVariable = new CollectedVariable("TESTIDVAR", List.of(new String[]{"V1", "V2"}), "TESTIDLOOP", "TESTIDPARENT");
-        collectedVariableList.add(collectedVariable);
-
-        SurveyUnitModel recentDTO = SurveyUnitModel.builder()
-                .idCampaign(idCampaign)
-                .mode(Mode.WEB)
-                .idUE(defaultIdUE)
-                .idQuest(idQuestionnaire)
-                .state(DataState.COLLECTED)
-                .fileDate(fileDate)
-                .recordDate(recordDate)
-                .externalVariables(externalVariableList)
-                .collectedVariables(collectedVariableList)
-                .build();
-        surveyUnitPersistencePortStub.getMongoStub().add(recentDTO);
-    }
-
-
-
-    private void addAdditionalDtoToMongoStub(DataState state,
-                                              String collectedVariableValue,
-                                              String externalVariableValue,
-                                              LocalDateTime fileDate,
-                                              LocalDateTime recordDate) {
-        List<Variable> externalVariableList = new ArrayList<>();
-        Variable variable = Variable.builder().idVar("TESTIDVAR").values(List.of(new String[]{externalVariableValue})).build();
-        externalVariableList.add(variable);
-
-        List<CollectedVariable> collectedVariableList = new ArrayList<>();
-        CollectedVariable collectedVariable = new CollectedVariable("TESTIDVAR", List.of(new String[]{collectedVariableValue}), "TESTIDLOOP", "TESTIDPARENT");
-        collectedVariableList.add(collectedVariable);
-
-        SurveyUnitModel recentDTO = SurveyUnitModel.builder()
-                .idCampaign("TESTIDCAMPAIGN")
-                .mode(Mode.WEB)
-                .idUE(defaultIdUE)
-                .idQuest(defaultIdQuest)
-                .state(state)
-                .fileDate(fileDate)
-                .recordDate(recordDate)
-                .externalVariables(externalVariableList)
-                .collectedVariables(collectedVariableList)
-                .build();
-        surveyUnitPersistencePortStub.getMongoStub().add(recentDTO);
-    }
-}
