+++ src/test/java/fr/insee/genesis/controller/rest/responses/ResponseControllerTest.java
@@ -1,0 +1,305 @@
+package fr.insee.genesis.controller.rest.responses;
+
+import cucumber.TestConstants;
+import fr.insee.genesis.Constants;
+import fr.insee.genesis.controller.dto.SurveyUnitDto;
+import fr.insee.genesis.controller.dto.SurveyUnitId;
+import fr.insee.genesis.controller.dto.SurveyUnitSimplified;
+import fr.insee.genesis.controller.utils.ControllerUtils;
+import fr.insee.genesis.domain.model.surveyunit.DataState;
+import fr.insee.genesis.domain.model.surveyunit.Mode;
+import fr.insee.genesis.domain.model.surveyunit.SurveyUnitModel;
+import fr.insee.genesis.domain.ports.api.SurveyUnitApiPort;
+import fr.insee.genesis.domain.service.surveyunit.SurveyUnitQualityService;
+import fr.insee.genesis.domain.service.surveyunit.SurveyUnitService;
+import fr.insee.genesis.infrastructure.utils.FileUtils;
+import fr.insee.genesis.stubs.ConfigStub;
+import fr.insee.genesis.stubs.SurveyUnitPersistencePortStub;
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.http.ResponseEntity;
+import org.springframework.util.FileSystemUtils;
+
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.time.LocalDateTime;
+import java.time.Month;
+import java.util.ArrayList;
+import java.util.List;
+
+class ResponseControllerTest {
+    //Given
+    static ResponseController responseControllerStatic;
+    static SurveyUnitPersistencePortStub surveyUnitPersistencePortStub;
+
+    static List<SurveyUnitId> surveyUnitIdList;
+    //Constants
+    static final String defaultIdUE = "TESTIDUE";
+    static final String defaultIdQuest = "TESTIDQUESTIONNAIRE";
+
+    @BeforeAll
+    static void init() {
+        surveyUnitPersistencePortStub = new SurveyUnitPersistencePortStub();
+        SurveyUnitApiPort surveyUnitApiPort = new SurveyUnitService(surveyUnitPersistencePortStub);
+
+        FileUtils fileUtils = new FileUtils(new ConfigStub());
+        responseControllerStatic = new ResponseController(
+                surveyUnitApiPort
+                , new SurveyUnitQualityService()
+                , fileUtils
+                , new ControllerUtils(fileUtils)
+        );
+
+        surveyUnitIdList = new ArrayList<>();
+        surveyUnitIdList.add(new SurveyUnitId(defaultIdUE));
+    }
+
+    @BeforeEach
+    void reset() throws IOException {
+        Utils.reset(surveyUnitPersistencePortStub);
+    }
+
+
+    //When + Then
+    @Test
+    void saveResponseFromXMLFileTest() throws Exception {
+        responseControllerStatic.saveResponsesFromXmlFile(
+                Path.of(TestConstants.TEST_RESOURCES_DIRECTORY, "IN/WEB/SAMPLETEST-PARADATA-v1/reponse-platine/data.complete.validated.STPDv1.20231122164209.xml").toString()
+                , Path.of(TestConstants.TEST_RESOURCES_DIRECTORY, "specs/SAMPLETEST-PARADATA-v1/ddi-SAMPLETEST-PARADATA-v1.xml").toString()
+                , Mode.WEB
+                , true
+        );
+
+        Assertions.assertThat(surveyUnitPersistencePortStub.getMongoStub()).isNotEmpty();
+    }
+
+    @Test
+    void saveOneFileNoCollected_NoNullPointerException(){
+        Assertions.assertThatCode(() -> responseControllerStatic.saveResponsesFromXmlFile(
+                Path.of(TestConstants.TEST_RESOURCES_DIRECTORY, "IN/WEB/SAMPLETEST-NO-COLLECTED/differential/data/data_diff_no_collected.xml").toString()
+                , Path.of(TestConstants.TEST_RESOURCES_DIRECTORY, "specs/SAMPLETEST-NO-COLLECTED/WEB/ddi_response_simple.xml").toString()
+                , Mode.WEB
+                , true
+        )).doesNotThrowAnyException();
+
+    }
+
+    @Test
+    void saveResponsesFromXmlCampaignFolderTest() throws Exception {
+        responseControllerStatic.saveResponsesFromXmlCampaignFolder(
+                "SAMPLETEST-PARADATA-v1"
+                , Mode.WEB
+                , true
+        );
+
+        Assertions.assertThat(surveyUnitPersistencePortStub.getMongoStub()).isNotEmpty();
+    }
+
+    @Test
+    void saveResponsesFromXmlCampaignFolderTest_noData() throws Exception {
+        surveyUnitPersistencePortStub.getMongoStub().clear();
+
+        responseControllerStatic.saveResponsesFromXmlCampaignFolder(
+                "TESTNODATA"
+                , Mode.WEB
+                , true
+        );
+
+        Assertions.assertThat(surveyUnitPersistencePortStub.getMongoStub()).isEmpty();
+    }
+
+    @Test
+    void saveResponsesFromAllCampaignFoldersTests(){
+        surveyUnitPersistencePortStub.getMongoStub().clear();
+        responseControllerStatic.saveResponsesFromAllCampaignFolders();
+
+        Assertions.assertThat(surveyUnitPersistencePortStub.getMongoStub()).isNotEmpty();
+    }
+
+    @Test
+    void findResponsesByUEAndQuestionnaireTest() {
+        ResponseEntity<List<SurveyUnitModel>> response = responseControllerStatic.findResponsesByUEAndQuestionnaire(defaultIdUE, defaultIdQuest);
+
+        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
+        Assertions.assertThat(response.getBody()).isNotNull().isNotEmpty();
+        Assertions.assertThat(response.getBody().getFirst().getIdUE()).isEqualTo(defaultIdUE);
+        Assertions.assertThat(response.getBody().getFirst().getIdQuest()).isEqualTo(defaultIdQuest);
+    }
+
+    @Test
+    void findAllResponsesByQuestionnaireTest() {
+        Path path = Path.of(TestConstants.TEST_RESOURCES_DIRECTORY, "OUT", defaultIdQuest);
+        File dir = new File(String.valueOf(path));
+        FileSystemUtils.deleteRecursively(dir);
+
+        ResponseEntity<Path> response = responseControllerStatic.findAllResponsesByQuestionnaire(defaultIdQuest);
+
+        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
+        Assertions.assertThat(response.getBody()).isNotNull();
+        Assertions.assertThat(Files.exists(path)).isTrue();
+        File[] dir_contents = dir.listFiles();
+        Assertions.assertThat(dir_contents).hasSize(1);
+        Assertions.assertThat(dir_contents[0].length()).isPositive().isNotNull();
+        FileSystemUtils.deleteRecursively(dir);
+        dir.deleteOnExit();
+    }
+
+    @Test
+    void getAllResponsesByQuestionnaireTestSequential() throws IOException {
+        //Given
+        surveyUnitPersistencePortStub.getMongoStub().clear();
+
+        for (int i = 0; i < Constants.BATCH_SIZE + 2; i++) {
+            Utils.addAdditionalDtoToMongoStub("TESTIDCAMPAIGN", defaultIdUE + i,
+                    LocalDateTime.of(2023, 1, 1, 0, 0, 0),
+                    LocalDateTime.of(2024, 1, 1, 0, 0, 0),
+                    surveyUnitPersistencePortStub);
+        }
+
+        //When
+        ResponseEntity<Path> response = responseControllerStatic.findAllResponsesByQuestionnaire(defaultIdQuest);
+
+        //Then
+        Assertions.assertThat(response).isNotNull();
+        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
+        Assertions.assertThat(response.getBody()).isNotNull();
+
+        Assertions.assertThat(response.getBody()).isNotNull();
+        Assertions.assertThat(response.getBody().toFile()).isNotNull().exists();
+
+        Files.deleteIfExists(response.getBody());
+    }
+
+    @Test
+    void getLatestByUETest() {
+        Utils.addAdditionalDtoToMongoStub(surveyUnitPersistencePortStub);
+
+        ResponseEntity<List<SurveyUnitModel>> response = responseControllerStatic.getLatestByUE(defaultIdUE, defaultIdQuest);
+
+        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
+        Assertions.assertThat(response.getBody()).isNotNull().isNotEmpty();
+        Assertions.assertThat(response.getBody().getFirst().getIdUE()).isEqualTo(defaultIdUE);
+        Assertions.assertThat(response.getBody().getFirst().getIdQuest()).isEqualTo(defaultIdQuest);
+        Assertions.assertThat(response.getBody().getFirst().getFileDate()).hasMonth(Month.FEBRUARY);
+    }
+
+    @Test
+    void getLatestByUEOneObjectTest() {
+        ResponseEntity<SurveyUnitSimplified> response = responseControllerStatic.getLatestByUEOneObject(defaultIdUE, defaultIdQuest, Mode.WEB);
+
+        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
+        Assertions.assertThat(response.getBody()).isNotNull();
+        Assertions.assertThat(response.getBody().getIdUE()).isEqualTo(defaultIdUE);
+        Assertions.assertThat(response.getBody().getIdQuest()).isEqualTo(defaultIdQuest);
+    }
+
+    @Test
+    void getLatestForUEListTest() {
+        ResponseEntity<List<SurveyUnitSimplified>> response = responseControllerStatic.getLatestForUEList(defaultIdQuest, surveyUnitIdList);
+
+        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
+        Assertions.assertThat(response.getBody()).isNotNull().isNotEmpty();
+        Assertions.assertThat(response.getBody().getFirst().getIdUE()).isEqualTo(defaultIdUE);
+    }
+
+    // Perret tests
+    @Test
+    void getLatestByStatesSurveyDataTest(){
+        //GIVEN
+        //Recent Collected already in stub
+        //Old Collected
+        Utils.addAdditionalDtoToMongoStub(DataState.COLLECTED,
+                "C OLD C", //<Collected/External> <NEW or OLD> <Collected/Edited>
+                "E OLD C",
+                LocalDateTime.of(1999,2,2,0,0,0),
+                LocalDateTime.of(1999,2,2,0,0,0),
+                surveyUnitPersistencePortStub
+        );
+
+        //Recent Edited
+        Utils.addAdditionalDtoToMongoStub(DataState.EDITED,
+                "C NEW E",
+                "E NEW E",
+                LocalDateTime.of(2025,2,2,0,0,0),
+                LocalDateTime.of(2025,2,2,0,0,0),
+                surveyUnitPersistencePortStub
+        );
+
+        //Old Edited
+        Utils.addAdditionalDtoToMongoStub(DataState.EDITED,
+                "C OLD E",
+                "E OLD E",
+                LocalDateTime.of(1999,2,2,0,0,0),
+                LocalDateTime.of(1999,2,2,0,0,0),
+                surveyUnitPersistencePortStub
+        );
+
+
+        //WHEN
+        ResponseEntity<SurveyUnitDto> response = responseControllerStatic.findResponsesByUEAndQuestionnaireLatestStates(
+                defaultIdUE,
+                defaultIdQuest
+        );
+
+
+        //THEN
+        SurveyUnitDto surveyUnitDto = response.getBody();
+        Assertions.assertThat(surveyUnitDto).isNotNull();
+
+        Assertions.assertThat(surveyUnitDto.getSurveyUnitId()).isEqualTo(defaultIdUE);
+
+        Assertions.assertThat(surveyUnitDto.getCollectedVariables().getFirst().getVariableName())
+                .isEqualTo("TESTIDVAR");
+
+        Assertions.assertThat(surveyUnitDto.getCollectedVariables().getFirst().getVariableStateDtoList()
+                        .stream().filter(
+                                variableStatePerret -> variableStatePerret.getState().equals(DataState.COLLECTED)
+                        ).toList().getFirst().getValue())
+                .isEqualTo("V1");
+        Assertions.assertThat(surveyUnitDto.getCollectedVariables().getFirst().getVariableStateDtoList()
+                        .stream().filter(
+                                variableStatePerret -> variableStatePerret.getState().equals(DataState.EDITED)
+                        ).toList().getFirst().getValue())
+                .isEqualTo("C NEW E");
+        Assertions.assertThat(surveyUnitDto.getCollectedVariables().getFirst().getVariableStateDtoList()
+                        .stream().filter(
+                                variableStatePerret -> variableStatePerret.getState().equals(DataState.COLLECTED)
+                        ).toList().getFirst().isActive())
+                .isFalse();
+        Assertions.assertThat(surveyUnitDto.getCollectedVariables().getFirst().getVariableStateDtoList()
+                        .stream().filter(
+                                variableStatePerret -> variableStatePerret.getState().equals(DataState.EDITED)
+                        ).toList().getFirst().isActive())
+                .isTrue();
+
+        Assertions.assertThat(surveyUnitDto.getExternalVariables().getFirst().getVariableName())
+                .isEqualTo("TESTIDVAR");
+        Assertions.assertThat(surveyUnitDto.getExternalVariables().getFirst().getVariableStateDtoList()
+                        .stream().filter(
+                                variableStatePerret -> variableStatePerret.getState().equals(DataState.COLLECTED)
+                        ).toList().getFirst().getValue())
+                .isEqualTo("V1");
+        Assertions.assertThat(surveyUnitDto.getExternalVariables().getFirst().getVariableStateDtoList()
+                        .stream().filter(
+                                variableStatePerret -> variableStatePerret.getState().equals(DataState.EDITED)
+                        ).toList().getFirst().getValue())
+                .isEqualTo("E NEW E");
+        Assertions.assertThat(surveyUnitDto.getExternalVariables().getFirst().getVariableStateDtoList()
+                        .stream().filter(
+                                variableStatePerret -> variableStatePerret.getState().equals(DataState.COLLECTED)
+                        ).toList().getFirst().isActive())
+                .isFalse();
+        Assertions.assertThat(surveyUnitDto.getExternalVariables().getFirst().getVariableStateDtoList()
+                        .stream().filter(
+                                variableStatePerret -> variableStatePerret.getState().equals(DataState.EDITED)
+                        ).toList().getFirst().isActive())
+                .isTrue();
+    }
+
+
+
+}
